Move winapi wrappers to sys module
